@@ -6,32 +6,31 @@ using Example1.Enums;
 using Example1.Model;
 using Newtonsoft.Json;
 
-namespace Example1.Command
+namespace Example1.Command;
+
+public class ImportPokemonCommand
 {
-    public class ImportPokemonCommand
+    public async Task<ImportingStatus> ImportPokemon(string filePath)
     {
-        public async Task<ImportingStatus> ImportPokemon(string filePath)
+        try
         {
-            try
-            {
-                Console.WriteLine($"Received pokemon to import: {filePath}..."); // usage of static method
+            Console.WriteLine($"Received pokemon to import: {filePath}...");
 
-                var fileContent = File.ReadAllText(filePath);
-                Pokemon pokemon = JsonConvert.DeserializeObject<Pokemon>(fileContent); // usage of framework method
-                pokemon.Timestamp = DateTime.Now;
+            var fileContent = await File.ReadAllTextAsync(filePath);
+            Pokemon pokemon = JsonConvert.DeserializeObject<Pokemon>(fileContent);
+            pokemon.Timestamp = DateTime.Now;
 
-                var database = new PokemonStore(); // new keyword
-                await database.SavePokemon(pokemon);
+            var database = new PokemonStore();
+            await database.SavePokemon(pokemon);
 
-                Console.WriteLine($"Pokemon saved. Id: {pokemon.Id}, Name: {pokemon.Name}, Type: {pokemon.Type}, Timestamp: {pokemon.Timestamp}");
+            Console.WriteLine($"Pokemon saved. Id: {pokemon.Id}, Name: {pokemon.Name}, Type: {pokemon.Type}, Timestamp: {pokemon.Timestamp}");
 
-                return ImportingStatus.Success;
-            }
-            catch (Exception e)
-            {
-                Console.WriteLine("Error! " + e.Message);
-                return ImportingStatus.Error;
-            }
+            return ImportingStatus.Success;
+        }
+        catch (Exception e)
+        {
+            Console.WriteLine("Error! " + e.Message);
+            return ImportingStatus.Error;
         }
     }
 }
