@@ -2,21 +2,16 @@
 using System.Threading.Tasks;
 using Example2.Command;
 
-namespace Example2
+namespace Example2;
+
+internal static class Program
 {
-    /*
-     * Created new class EntityReader for isolation
-     * Still tightly coupled
-     */
-    internal static class Program
+    static async Task Main()
     {
-        static async Task Main(string[] args)
-        {
-            var command = new ImportPokemonCommand();
+        var command = new ImportPokemonCommand();
 
-            await command.ImportPokemon(@"App_Data\data.json");
+        await command.ImportPokemon(@"App_Data\data.json");
 
-            Console.ReadKey();
-        }
+        Console.ReadKey();
     }
 }
