@@ -1,17 +1,20 @@
 ﻿using System;
+using System.Text.Json;
 using Example2.Model;
-using Newtonsoft.Json;
 
-namespace Example2.Reader
+namespace Example2.Reader;
+
+public class PokemonReader
 {
-    public class PokemonReader
+    public Pokemon ReadPokemon(string fileContent)
     {
-        public Pokemon ReadPokemon(string fileContent)
+        var pokemon = JsonSerializer.Deserialize<Pokemon>(fileContent, new JsonSerializerOptions
         {
-            Pokemon pokemon = JsonConvert.DeserializeObject<Pokemon>(fileContent);
-            pokemon.Timestamp = DateTime.Now;
-            Console.WriteLine($"Pokemon deserialized. Id: {pokemon.Id}, Name: {pokemon.Name}, Type: {pokemon.Type}, Timestamp: {pokemon.Timestamp}");
-            return pokemon;
-        }
+            PropertyNameCaseInsensitive = true,
+        }); 
+
+        pokemon.Timestamp = DateTime.Now;
+        Console.WriteLine($"Pokemon deserialized. Id: {pokemon.Id}, Name: {pokemon.Name}, Type: {pokemon.Type}, Timestamp: {pokemon.Timestamp}");
+        return pokemon;
     }
 }
