@@ -1,4 +1,5 @@
 ﻿using System;
+using System.IO;
 using System.Threading.Tasks;
 using Example3.Command;
 using Example3.Db;
@@ -11,144 +12,142 @@ using FluentAssertions;
 using Moq;
 using NUnit.Framework;
 
-namespace Example3_Tests
+namespace Example3_Tests;
+
+[TestFixture]
+public class Example3Tests
 {
-    /*
-     * Tests no longer require to write physical file to hard drive.
-     * Instead mocks can be set up to return what we need.
-     */
-    [TestFixture]
-    public class Example3Tests
+    private Mock<IDateTimeWrapper> _dateTimeWrapperMock;
+    private Mock<IPokemonStore> _databaseMock;
+    private Mock<IPokemonReader> _pokemonReaderMock;
+    private Mock<ILogger> _loggerMock;
+    private Mock<IFileSystemWrapper> _fileSystemWrapperMock;
+
+    private FileInfo _testData;
+    private Pokemon _expectedPokemon;
+
+    [SetUp]
+    public void SetUp()
     {
-        private Mock<IDateTimeWrapper> _dateTimeWrapperMock;
-        private Mock<IPokemonStore> _databaseMock;
-        private Mock<IPokemonReader> _pokemonReaderMock;
-        private Mock<ILogger> _loggerMock;
-        private Mock<IFileSystemWrapper> _fileSystemWrapperMock;
+        _dateTimeWrapperMock = new Mock<IDateTimeWrapper>();
+        _databaseMock = new Mock<IPokemonStore>();
+        _pokemonReaderMock = new Mock<IPokemonReader>();
+        _loggerMock = new Mock<ILogger>();
+        _fileSystemWrapperMock = new Mock<IFileSystemWrapper>();
 
-        private string _json;
-        private Pokemon _expectedPokemon;
-
-        private const string FakeFilePath = "test.json";
-
-        [SetUp]
-        public void SetUp()
+        _testData = new FileInfo(@"App_Data\testdata.json");
+        _expectedPokemon = new Pokemon
         {
-            _dateTimeWrapperMock = new Mock<IDateTimeWrapper>();
-            _databaseMock = new Mock<IPokemonStore>();
-            _pokemonReaderMock = new Mock<IPokemonReader>();
-            _loggerMock = new Mock<ILogger>();
-            _fileSystemWrapperMock = new Mock<IFileSystemWrapper>();
+            Id = 4,
+            Name = "Charmander",
+            Type = PokemonType.Fire,
+            Timestamp = new DateTime(2010, 1, 1)
+        };
+    }
 
-            _json = @"{ ""id"": 1, ""name"": ""Charmander"", ""type"": 1 }";
-            _expectedPokemon = new Pokemon
-            {
-                Id = 1,
-                Name = "Charmander",
-                Type = PokemonType.Fire,
-                Timestamp = new DateTime(2010, 1, 1)
-            };
-        }
+    [Test]
+    public async Task PokemonReader_ShouldDeserializePokemon()
+    {
+        // Arrange
+        _dateTimeWrapperMock.Setup(x => x.GetNow()).Returns(new DateTime(2010, 1, 1));
+        var stream = File.OpenRead(_testData.FullName);
+        var reader = new PokemonReader(_dateTimeWrapperMock.Object, _loggerMock.Object);
 
-        [Test]
-        public void PokemonReader_ShouldDeserializePokemon()
-        {
-            // Arrange
-            _dateTimeWrapperMock.Setup(x => x.GetNow()).Returns(new DateTime(2010, 1, 1));
-            var reader = new PokemonReader(_dateTimeWrapperMock.Object, _loggerMock.Object);
+        // Act
+        var result = await reader.ReadPokemon(stream);
 
-            // Act
-            var result = reader.ReadPokemon(_json);
+        // Assert
+        result.Should().BeEquivalentTo(_expectedPokemon);
+    }
 
-            // Assert
-            result.Should().BeEquivalentTo(_expectedPokemon);
-        }
+    [Test]
+    public async Task PokemonReader_ShouldLogDeserializationOfPokemon()
+    {
+        // Arrange
+        _dateTimeWrapperMock.Setup(x => x.GetNow()).Returns(new DateTime(2010, 1, 1));
+        var stream = File.OpenRead(_testData.FullName);
+        var reader = new PokemonReader(_dateTimeWrapperMock.Object, _loggerMock.Object);
 
-        [Test]
-        public void PokemonReader_ShouldLogDeserializationOfPokemon()
-        {
-            // Arrange
-            _dateTimeWrapperMock.Setup(x => x.GetNow()).Returns(new DateTime(2010, 1, 1));
-            var reader = new PokemonReader(_dateTimeWrapperMock.Object, _loggerMock.Object);
+        // Act
+        await reader.ReadPokemon(stream);
 
-            // Act
-            reader.ReadPokemon(_json);
+        // Assert
+        _loggerMock.Verify(x => x.Information($"Pokemon deserialized. Id: {_expectedPokemon.Id}, Name: {_expectedPokemon.Name}, Type: {_expectedPokemon.Type}, Timestamp: {_expectedPokemon.Timestamp}"));
+    }
 
-            // Assert
-            _loggerMock.Verify(x => x.Information($"Pokemon deserialized. Id: {_expectedPokemon.Id}, Name: {_expectedPokemon.Name}, Type: {_expectedPokemon.Type}, Timestamp: {_expectedPokemon.Timestamp}"));
-        }
+    [Test]
+    public async Task ImportPokemonCommand_ShouldLogReceivedFile()
+    {
+        // Arrange
+        var command = new ImportPokemonCommand(_loggerMock.Object, _databaseMock.Object, _pokemonReaderMock.Object, _fileSystemWrapperMock.Object);
 
-        [Test]
-        public async Task ImportPokemonCommand_ShouldLogReceivedFile()
-        {
-            // Arrange
-            var command = new ImportPokemonCommand( _loggerMock.Object, _databaseMock.Object, _pokemonReaderMock.Object, _fileSystemWrapperMock.Object);
+        // Act
+        await command.ImportPokemon(_testData.FullName);
 
-            // Act
-            await command.ImportPokemon(FakeFilePath);
+        // Assert
+        _loggerMock.Verify(x => x.Information($"Received pokemon to import: {_testData.FullName}..."));
+    }
 
-            // Assert
-            _loggerMock.Verify(x => x.Information($"Received pokemon to import: {FakeFilePath}..."));
-        }
+    [Test]
+    public async Task ImportPokemonCommand_ShouldReadFile()
+    {
+        // Arrange
+        var command = new ImportPokemonCommand(_loggerMock.Object, _databaseMock.Object, _pokemonReaderMock.Object, _fileSystemWrapperMock.Object);
 
-        [Test]
-        public async Task ImportPokemonCommand_ShouldReadFile()
-        {
-            // Arrange
-            var command = new ImportPokemonCommand( _loggerMock.Object, _databaseMock.Object, _pokemonReaderMock.Object, _fileSystemWrapperMock.Object);
+        // Act
+        await command.ImportPokemon(_testData.FullName);
 
-            // Act
-            await command.ImportPokemon(FakeFilePath);
+        // Assert
+        _fileSystemWrapperMock.Verify(x => x.OpenRead(_testData.FullName));
+    }
 
-            // Assert
-            _fileSystemWrapperMock.Verify(x => x.ReadFile(FakeFilePath));
-        }
+    [Test]
+    public async Task ImportPokemonCommand_ShouldReadPokemon()
+    {
+        // Arrange
+        const string fakeFilePath = "fake.json";
+        var stream = File.OpenRead(_testData.FullName);
+        _fileSystemWrapperMock.Setup(x => x.OpenRead(fakeFilePath)).Returns(stream);
+        var command = new ImportPokemonCommand(_loggerMock.Object, _databaseMock.Object, _pokemonReaderMock.Object, _fileSystemWrapperMock.Object);
 
-        [Test]
-        public async Task ImportPokemonCommand_ShouldReadPokemon()
-        {
-            // Arrange
-            var command = new ImportPokemonCommand( _loggerMock.Object, _databaseMock.Object, _pokemonReaderMock.Object, _fileSystemWrapperMock.Object);
+        // Act
+        await command.ImportPokemon(fakeFilePath);
 
-            _fileSystemWrapperMock.Setup(x => x.ReadFile(FakeFilePath)).Returns(_json);
+        // Assert
+        _pokemonReaderMock.Verify(x => x.ReadPokemon(stream));
+    }
 
-            // Act
-            await command.ImportPokemon(FakeFilePath);
+    [Test]
+    public async Task ImportPokemonCommand_ShouldSaveDatabase()
+    {
+        // Arrange
+        const string fakeFilePath = "fake.json";
+        var stream = File.OpenRead(_testData.FullName);
+        _fileSystemWrapperMock.Setup(x => x.OpenRead(fakeFilePath)).Returns(stream);
+        _pokemonReaderMock.Setup(x => x.ReadPokemon(stream)).ReturnsAsync(_expectedPokemon);
+        var command = new ImportPokemonCommand(_loggerMock.Object, _databaseMock.Object, _pokemonReaderMock.Object, _fileSystemWrapperMock.Object);
 
-            // Assert
-            _pokemonReaderMock.Verify(x => x.ReadPokemon(_json));
-        }
+        // Act
+        await command.ImportPokemon(fakeFilePath);
 
-        [Test]
-        public async Task ImportPokemonCommand_ShouldSaveDatabase()
-        {
-            // Arrange
-            var command = new ImportPokemonCommand( _loggerMock.Object, _databaseMock.Object, _pokemonReaderMock.Object, _fileSystemWrapperMock.Object);
+        // Assert
+        _databaseMock.Verify(x => x.SavePokemon(_expectedPokemon));
+    }
 
-            _fileSystemWrapperMock.Setup(x => x.ReadFile(FakeFilePath)).Returns(_json);
-            _pokemonReaderMock.Setup(x => x.ReadPokemon(_json)).Returns(_expectedPokemon);
+    [Test]
+    public async Task ImportPokemonCommand_ShouldReturnImportingStatusSuccess()
+    {
+        // Arrange
+        const string fakeFilePath = "fake.json";
+        var stream = File.OpenRead(_testData.FullName);
+        _fileSystemWrapperMock.Setup(x => x.OpenRead(fakeFilePath)).Returns(stream);
+        _pokemonReaderMock.Setup(x => x.ReadPokemon(stream)).ReturnsAsync(_expectedPokemon);
+        var command = new ImportPokemonCommand(_loggerMock.Object, _databaseMock.Object, _pokemonReaderMock.Object, _fileSystemWrapperMock.Object);
 
-            // Act
-            await command.ImportPokemon(FakeFilePath);
+        // Act
+        var result = await command.ImportPokemon(fakeFilePath);
 
-            // Assert
-            _databaseMock.Verify(x => x.SavePokemon(_expectedPokemon));
-        }
-
-        [Test]
-        public async Task ImportPokemonCommand_ShouldReturnImportingStatusSuccess()
-        {
-            // Arrange
-            var command = new ImportPokemonCommand( _loggerMock.Object, _databaseMock.Object, _pokemonReaderMock.Object, _fileSystemWrapperMock.Object);
-
-            _fileSystemWrapperMock.Setup(x => x.ReadFile(FakeFilePath)).Returns(_json);
-            _pokemonReaderMock.Setup(x => x.ReadPokemon(_json)).Returns(_expectedPokemon);
-
-            // Act
-            var result = await command.ImportPokemon(FakeFilePath);
-
-            // Assert
-            Assert.AreEqual(ImportingStatus.Success, result);
-        }
+        // Assert
+        Assert.AreEqual(ImportingStatus.Success, result);
     }
 }
