@@ -1,9 +1,8 @@
 ﻿using System;
 
-namespace Example4.Wrappers
+namespace Example4.Wrappers;
+
+public interface IDateTimeWrapper
 {
-    public interface IDateTimeWrapper
-    {
-        DateTime GetNow();
-    }
+    DateTime GetNow();
 }
