@@ -2,20 +2,14 @@
 using System.Threading.Tasks;
 using Example5.Model;
 
-namespace Example5.Db
+namespace Example5.Db;
+
+public class PokemonStore : IPokemonStore
 {
-    public class PokemonStore : IPokemonStore
+    public async Task SavePokemon(Pokemon pokemon)
     {
-        public async Task SavePokemon(Pokemon pokemon)
-        {
-            Console.WriteLine("Saving to database: " + pokemon.Name);
+        Console.WriteLine("Saving to database: " + pokemon.Name);
 
-            await Task.Delay(0);
-        }
-
-        public bool IsValidationEnabled()
-        {
-            return true;
-        }
+        await Task.Delay(0);
     }
 }
