@@ -1,16 +1,15 @@
 ﻿using System;
 using Example5.Enums;
 
-namespace Example5.Model
+namespace Example5.Model;
+
+public class Pokemon
 {
-    public class Pokemon
-    {
-        public int Id { get; set; }
+    public int Id { get; set; }
 
-        public string Name { get; set; }
+    public string Name { get; set; }
 
-        public PokemonType Type { get; set; }
+    public PokemonType Type { get; set; }
 
-        public DateTime? Timestamp { get; set; }
-    }
+    public DateTime? Timestamp { get; set; }
 }
