@@ -1,12 +1,11 @@
 ﻿using System;
 
-namespace Example5.Wrappers
+namespace Example5.Wrappers;
+
+public class DateTimeWrapper : IDateTimeWrapper
 {
-    public class DateTimeWrapper : IDateTimeWrapper
+    public DateTime GetNow()
     {
-        public DateTime GetNow()
-        {
-            return DateTime.Now;
-        }
+        return DateTime.Now;
     }
 }
